+++ src/kodi/generator.py
@@ -27,14 +27,15 @@ from kodi.io_utils import determine_dirs, read_id, skip, proceed
 logger = logging.getLogger("kodi.generator")
 
 
-def generate(dir, idtype="imdb", recursive=True, pattern="*.imdb", delay=1, dry_run=False, overwrite=False,
+def generate(path, idtype="imdb", recursive=True, pattern="*.imdb", delay=1, dry_run=False, overwrite=False,
              language="en", fanart="none", fanart_file="folder.jpg", interactive=False, episodes=False,
+             episode_pattern="*S??E??*.*", season_group=".*S([0-9]?[0-9])E.*", episode_group=".*E([0-9]?[0-9]).*",
              ua="Mozilla"):
     """
     Traverses the directory and generates the .nfo files.
 
-    :param dir: the directory to traverse
-    :type dir: str
+    :param path: the directory to traverse
+    :type path: str
     :param idtype: how to interpret the ID files (choices: 'imdb')
     :type idtype: str
     :param recursive: whether to search recursively
@@ -57,12 +58,18 @@ def generate(dir, idtype="imdb", recursive=True, pattern="*.imdb", delay=1, dry_
     :type interactive: bool
     :param episodes: whether to generate episode information as well
     :type episodes: bool
+    :param episode_pattern: the pattern to use for locating episode files
+    :type episode_pattern: str
+    :param season_group: the regular expression to extract the season (first group)
+    :type season_group: str
+    :param episode_group: the regular expression to extract the episode (first group)
+    :type episode_group: str
     :param ua: User agent for requests
     :type ua: str
     """
 
     dirs = []
-    determine_dirs(dir, recursive, dirs)
+    determine_dirs(path, recursive, dirs)
     dirs.sort()
     logger.info("# dirs: %d" % len(dirs))
     if interactive:
@@ -89,8 +96,10 @@ def generate(dir, idtype="imdb", recursive=True, pattern="*.imdb", delay=1, dry_
             try:
                 if idtype == "imdb":
                     file_generated = generate_imdb(id, language=language, fanart=fanart, fanart_file=fanart_file,
-                                                   episodes=episodes, path=d, overwrite=overwrite,
-                                                   dry_run=dry_run, ua=ua)
+                                                   path=d, overwrite=overwrite, dry_run=dry_run,
+                                                   episodes=episodes, episode_pattern=episode_pattern,
+                                                   season_group=season_group, episode_group=episode_group,
+                                                   ua=ua)
                 else:
                     logger.critical("Unhandled ID type: %s" % idtype)
                     return
@@ -125,6 +134,9 @@ def main(args=None):
     parser.add_argument("--fanart", dest="fanart", choices=["none", "download", "download-missing", "use-existing"], default="none", required=False, help="how to deal with fan-art")
     parser.add_argument("--fanart_file", metavar="FILE", dest="fanart_file", default="folder.jpg", required=False, help="when downloading or using existing fanart, use this filename")
     parser.add_argument("--episodes", action="store_true", dest="episodes", required=False, help="whether to generate .nfo files for episodes as well")
+    parser.add_argument("--episode_pattern", dest="episode_pattern", required=False, default="*S??E??*.*", help="the shell pattern to use for locating episode files")
+    parser.add_argument("--season_group", dest="season_group", required=False, default=".*S([0-9]?[0-9])E.*", help="the regular expression to extract the season (first group)")
+    parser.add_argument("--episode_group", dest="episode_group", required=False, default=".*E([0-9]?[0-9]).*", help="the regular expression to extract the episode (first group)")
     parser.add_argument("--dry_run", action="store_true", dest="dry_run", required=False, help="whether to perform a 'dry-run', ie only outputting the .nfo content to stdout but not saving it to files")
     parser.add_argument("--overwrite", action="store_true", dest="overwrite", required=False, help="whether to overwrite existing .nfo files, ie recreating them with freshly retrieved data")
     parser.add_argument("--verbose", action="store_true", dest="verbose", required=False, help="whether to output logging information")
@@ -143,10 +155,12 @@ def main(args=None):
     logger.debug(parsed)
     if parsed.interactive:
         logger.info("Entering interactive mode")
-    generate(dir=parsed.dir, idtype=parsed.type, recursive=parsed.recursive, pattern=parsed.pattern,
+    generate(path=parsed.dir, idtype=parsed.type, recursive=parsed.recursive, pattern=parsed.pattern,
              dry_run=parsed.dry_run, overwrite=parsed.overwrite, language=parsed.language,
              fanart=parsed.fanart, fanart_file=parsed.fanart_file, interactive=parsed.interactive,
-             episodes=parsed.episodes, ua=parsed.user_agent)
+             episodes=parsed.episodes, episode_pattern=parsed.episode_pattern,
+             season_group=parsed.season_group, episode_group=parsed.episode_group,
+             ua=parsed.user_agent)
 
 
 def sys_main():